MEED-374 Rename snapshot component
