+++ deeds-dapp-webapp/src/main/webapp/vue-app/snapshot/initComponents.js
@@ -16,7 +16,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-import Snapshot from './components/Snapshot.vue';
+import Overview from './components/Overview.vue';
 import PriceChart from './components/PriceChart.vue';
 import TradeMeeds from './components/TradeMeeds.vue';
 import Assets from './components/Assets.vue';
@@ -24,7 +24,7 @@ import TokenAssets from './components/assets/TokenAssets.vue';
 import DeedAssets from './components/assets/DeedAssets.vue';
 
 const components = {
-  'deeds-snapshot': Snapshot,
+  'deeds-overview': Overview,
   'deeds-price-chart': PriceChart,
   'deeds-trade-meeds': TradeMeeds,
   'deeds-assets': Assets,
